TST: move threshold testing, add test for n jobs
