+++ borsar/tests/test_stats.py
@@ -5,7 +5,9 @@ import pandas as pd
 import statsmodels.api as sm
 from statsmodels.stats.anova import AnovaRM
 
-from borsar.stats import compute_regression_t, format_pvalue, _find_stat_fun
+from borsar.stats import (compute_regression_t, format_pvalue, _find_stat_fun,
+                          _compute_threshold_via_permutations,
+                          _compute_threshold)
 
 
 def test_compute_regression_t():
@@ -135,3 +137,65 @@ def test_find_stat_fun():
     f_val_good = res.anova_table.loc['rep', 'F Value']
 
     np.testing.assert_almost_equal(f_val, f_val_good)
+
+
+def test_compute_threshold_via_permutations():
+    """Make sure that threshold computed through permutations is correct.
+
+    Check that the threshold computed through permutations/randomization
+    on data that fulfills assumptions of analytical tests is sufficiently
+    close to the analytical threshold.
+    """
+    n_groups = 2
+
+    for paired in [False, True]:
+        if paired:
+            n_obs = [101, 101]
+            data = [np.random.randn(n_obs[0])]
+            data.append(data[0] + np.random.randn(n_obs[0]))
+        else:
+            n_obs = [102, 100]
+            data = [np.random.randn(n) for n in n_obs]
+
+        analytical_threshold = _compute_threshold(
+            data=data, threshold=None, p_threshold=0.05, paired=paired,
+            one_sample=False)
+
+        stat_fun = _find_stat_fun(
+            n_groups, paired=paired, tail='both')
+
+        permutation_threshold = (
+            _compute_threshold_via_permutations(
+                data, paired=paired, tail='both', stat_fun=stat_fun,
+                n_permutations=2_000, progress=False
+            )
+        )
+
+        avg_perm = np.abs(permutation_threshold).mean()
+        error = analytical_threshold - avg_perm
+
+        print('paired:', paired)
+        print('analytical_threshold:', analytical_threshold)
+        print('permutation threshold:', permutation_threshold)
+        print('average permutation threshold:', avg_perm)
+        print('difference:', error)
+
+        assert np.abs(error) < 0.15
+
+
+def test_compute_threshold_via_permutations_n_jobs():
+    data = [np.random.randn(12, 10, 10), np.random.randn(12, 10, 10)]
+    stat_fun = _find_stat_fun(2, paired=True, tail='both')
+    pos_thresh, neg_thresh = _compute_threshold_via_permutations(
+        data, paired=True, tail='both', stat_fun=stat_fun,
+        n_permutations=1_000, progress=False
+    )
+    pos_thresh_jobs, neg_thresh_jobs = _compute_threshold_via_permutations(
+        data, paired=True, tail='both', stat_fun=stat_fun,
+        n_permutations=1_000, progress=False, n_jobs=2
+    )
+
+    # most differences are < 0.5, but some are larger
+    # (independent permutation runs)
+    assert (np.abs(pos_thresh - pos_thresh_jobs) < 0.5).mean() > 0.9
+    assert (np.abs(neg_thresh - neg_thresh_jobs) < 0.5).mean() > 0.9
